Separate tcp from main
Dispatcher: Pass tcp_pcb as argument instead of using a variable
