+++ pico-client/waterer/CMakeLists.txt
@@ -1,3 +1,12 @@
+cmake_minimum_required(VERSION 3.13)
+
+set(PICO_BOARD pico_w)
+
+if (NOT TARGET pico_stdlib)
+  include(pico_sdk_import.cmake)
+  project(watering LANGUAGES C CXX ASM VERSION 0.0.1)
+endif()
+
 function(create_slot_target slot_name flash_offset)
   set(
     FLASH_ORIGIN ${flash_offset}
@@ -10,6 +19,7 @@ function(create_slot_target slot_name flash_offset)
     main.c
     dispatcher.c
     sched.c
+    tcp_client.c
     ../common/shared_mem.c
     ${CMAKE_SOURCE_DIR}/common/crc.c
     ${CMAKE_SOURCE_DIR}/../common/helpers.c
@@ -22,6 +32,7 @@ function(create_slot_target slot_name flash_offset)
     ${CMAKE_CURRENT_LIST_DIR}
     ${CMAKE_CURRENT_LIST_DIR}/common/
     ${CMAKE_CURRENT_LIST_DIR}/..
+    ${CMAKE_CURRENT_LIST_DIR}/../common/
   )
 
   target_link_libraries(${target_name}
@@ -31,8 +42,10 @@ function(create_slot_target slot_name flash_offset)
     hardware_adc
     hardware_i2c
     hardware_clocks
-    hardware_rosc
     hardware_irq
+    hardware_spi
+    hardware_adc
+    hardware_watchdog
   )
 
   # Create linker script with correct origin
@@ -44,10 +57,8 @@ function(create_slot_target slot_name flash_offset)
     @ONLY
   )
 
-  # Use the custom linker script
   pico_set_linker_script(${target_name} "${linker_script}")
 
-  # Add extra outputs
   pico_add_extra_outputs(${target_name})
 
   # Rename output files
